packages/core/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "loft-cli-core"
version = "0.6.3"
description = "Shared models, specs, and registry infrastructure for loft-cli"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["loft_cli_core*"]
